feature/1: Inversed_related_products static passed
