--- app/code/Study/AdvancedContactForm/Model/Config.php
@@ -1,10 +1,0 @@
-<?php
-
-
-namespace Study\AdvancedContactForm\Model;
-
-
-class Config
-{
-
-}
