--- app/code/Study/AdvancedContactForm/etc/adminhtml/routes.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0"?>
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:App/etc/routes.xsd">
-    <router id="admin">
-        <route id="uiform" frontName="uiform">
-            <module name="Study_AdvancedContactForm"/>
-        </route>
-    </router>
-</config>
