--- app/code/Study/AdvancedContactForm/etc/adminhtml/system.xml
@@ -1,14 +1,0 @@
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:module:Magento_Config:etc/system_file.xsd">
-    <system>
-            <section id="reports" translate="label" sortOrder="130" showInDefault="1" showInWebsite="1" showInStore="1">
-                <tab>general</tab>
-                <group id="options" translate="label" type="text" sortOrder="1" showInDefault="1">
-                    <field id="advanced_contacts_enabled" translate="label comment" type="select" sortOrder="1" showInDefault="1" canRestore="1">
-                        <label>Enable Advanced contact logging</label>
-                        <source_model>Magento\Config\Model\Config\Source\Yesno</source_model>
-                        <comment>If disabled, advanced contacts logging will be disabled</comment>
-                    </field>
-                </group>
-            </section>
-    </system>
-</config>
