--- app/code/Study/AdvancedContactForm/etc/di.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0"?>
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
-    <type name="Magento\Contact\Controller\Index\Post">
-        <plugin name="advanced_contact_post_plugin" type="Study\AdvancedContactForm\Plugin\Controller\Index\Post2" />
-    </type>
-    <type name="Magento\Framework\View\Element\UiComponent\DataProvider\CollectionFactory">
-        <arguments>
-            <argument name="collections" xsi:type="array">
-                <item name="advanced_contact_data_source" xsi:type="string">Study\AdvancedContactForm\Model\ResourceModel\Contact\Grid\Collection</item>
-            </argument>
-        </arguments>
-    </type>
-    <virtualType name="Study\AdvancedContactForm\Model\ResourceModel\Contact\Grid\Collection" type="Magento\Framework\View\Element\UiComponent\DataProvider\SearchResult">
-        <arguments>
-            <argument name="mainTable" xsi:type="string">study_advancedcontacts</argument>
-            <argument name="resourceModel" xsi:type="string">Study\AdvancedContactForm\Model\ResourceModel\Contact</argument>
-        </arguments>
-    </virtualType>
-</config>
