--- app/code/Study/AdvancedContactForm/registration.php
@@ -1,8 +1,0 @@
-<?php
-use Magento\Framework\Component\ComponentRegistrar;
-
-ComponentRegistrar::register(
-    ComponentRegistrar::MODULE,
-    'Study_AdvancedContactForm',
-    __DIR__
-);
