--- app/code/Study/AdvancedContactForm/view/adminhtml/layout/uiform_index_index.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0"?>
-<page xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
-    <update handle="styles"/>
-    <body>
-        <referenceContainer name="content">
-            <uiComponent name="advancedcontact_form_listing"/>
-        </referenceContainer>
-<!--        <referenceContainer name="content">-->
-<!--            <uiComponent name="dev_grid_category_listing"/>-->
-<!--        </referenceContainer>-->
-
-    </body>
-</page>
