--- app/code/Study/AdvancedContactForm/view/adminhtml/templates/adminpage.phtml
@@ -1,1 +1,0 @@
-<p>Hello World!</p>
