--- app/code/Study/AdvancedContactForm/view/adminhtml/ui_component/advancedcontact_form_listing.xml
@@ -1,80 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<listing xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:module:Magento_Ui:etc/ui_configuration.xsd">
-    <argument name="data" xsi:type="array">
-        <item name="js_config" xsi:type="array">
-            <item name="provider" xsi:type="string">advancedcontact_form_listing.advancedcontactform_listing_data_source</item>
-            <item name="deps" xsi:type="string">advancedcontact_form_listing.advancedcontactform_listing_data_source</item>
-        </item>
-        <item name="spinner" xsi:type="string">spinner_columns</item>
-    </argument>
-    <dataSource name="advancedcontactform_listing_data_source">
-        <argument name="dataProvider" xsi:type="configurableObject">
-            <argument name="class" xsi:type="string">Study\AdvancedContactForm\Ui\DataProvider\Contact\DataProvider</argument>
-            <argument name="name" xsi:type="string">advancedcontactform_listing_data_source</argument>
-            <argument name="primaryFieldName" xsi:type="string">contact_id</argument>
-            <argument name="requestFieldName" xsi:type="string">id</argument>
-            <argument name="data" xsi:type="array">
-                <item name="config" xsi:type="array">
-                    <item name="component" xsi:type="string">Magento_Ui/js/grid/provider</item>
-                    <item name="update_url" xsi:type="url" path="mui/index/render"/>
-                    <item name="storageConfig" xsi:type="array">
-                        <item name="indexField" xsi:type="string">contact_id</item>
-                    </item>
-                </item>
-            </argument>
-        </argument>
-    </dataSource>
-    <columns name="spinner_columns">
-    <column name="contact_id">
-        <argument name="data" xsi:type="array">
-            <item name="config" xsi:type="array">
-                <item name="filter" xsi:type="string">textRange</item>
-                <item name="sorting" xsi:type="string">asc</item>
-                <item name="label" xsi:type="string" translate="true">ID</item>
-            </item>
-        </argument>
-    </column>
-    <column name="name">
-        <argument name="data" xsi:type="array">
-            <item name="config" xsi:type="array">
-                <item name="filter" xsi:type="string">textRange</item>
-                <item name="label" xsi:type="string" translate="true">Name</item>
-            </item>
-        </argument>
-    </column>
-    <column name="email">
-        <argument name="data" xsi:type="array">
-            <item name="config" xsi:type="array">
-                <item name="filter" xsi:type="string">textRange</item>
-                <item name="label" xsi:type="string" translate="true">Email</item>
-            </item>
-        </argument>
-    </column>
-    <column name="telephone">
-        <argument name="data" xsi:type="array">
-            <item name="config" xsi:type="array">
-                <item name="filter" xsi:type="string">textRange</item>
-                <item name="label" xsi:type="string" translate="true">Phone</item>
-            </item>
-        </argument>
-    </column>
-    <column name="comment">
-        <argument name="data" xsi:type="array">
-            <item name="config" xsi:type="array">
-                <item name="filter" xsi:type="string">textRange</item>
-                <item name="label" xsi:type="string" translate="true">Comment</item>
-            </item>
-        </argument>
-    </column>
-    <column name="created_at" class="Magento\Ui\Component\Listing\Columns\Date">
-        <argument name="data" xsi:type="array">
-            <item name="config" xsi:type="array">
-                <item name="filter" xsi:type="string">dateRange</item>
-                <item name="component" xsi:type="string">Magento_Ui/js/grid/columns/date</item>
-                <item name="dataType" xsi:type="string">date</item>
-                <item name="label" xsi:type="string" translate="true">Created at : </item>
-            </item>
-        </argument>
-    </column>
-    </columns>
-</listing>
