+++ app/code/Study/ImprovedContact/Controller/Adminhtml/Index/Index.php
@@ -1,0 +1,24 @@
+<?php
+declare(strict_types=1);
+
+namespace Study\ImprovedContact\Controller\Adminhtml\Index;
+
+use Magento\Backend\App\Action;
+use Magento\Framework\App\Action\HttpGetActionInterface;
+use Magento\Framework\App\ResponseInterface as ResponseInterfaceAlias;
+use Magento\Framework\Controller\ResultFactory;
+use Magento\Framework\Controller\ResultInterface;
+use Magento\Framework\View\Result\Page as PageAlias;
+
+class Index extends Action implements HttpGetActionInterface
+{
+    /**
+     * Main method
+     *
+     * @return ResponseInterfaceAlias|ResultInterface|PageAlias
+     */
+    public function execute()
+    {
+        return $this->resultFactory->create(ResultFactory::TYPE_PAGE);
+    }
+}
