+++ app/code/Study/ImprovedContact/Ui/DataProvider/Contact/DataProvider.php
@@ -1,0 +1,42 @@
+<?php
+declare(strict_types=1);
+
+namespace Study\ImprovedContact\Ui\DataProvider\Contact;
+
+use Magento\Ui\DataProvider\AbstractDataProvider;
+use Study\ImprovedContact\Model\ResourceModel\AdvancedContactForm\CollectionFactory;
+
+class DataProvider extends AbstractDataProvider
+{
+    /**
+     * @var CollectionFactory
+     */
+    protected $collection;
+
+    /**
+     * @param string $name
+     * @param string $primaryFieldName
+     * @param string $requestFieldName
+     * @param CollectionFactory $colectionFactory
+     * @param array $meta
+     * @param array $data
+     * @SuppressWarnings(PHPMD.ExcessiveParameterList)
+     */
+    public function __construct(
+        $name,
+        $primaryFieldName,
+        $requestFieldName,
+        CollectionFactory $colectionFactory,
+        array $meta = [],
+        array $data = []
+    ) {
+        parent::__construct(
+            $name,
+            $primaryFieldName,
+            $requestFieldName,
+            $meta,
+            $data
+        );
+        $this->collection = $colectionFactory->create();
+    }
+}
