+++ app/code/Study/ImprovedContact/etc/adminhtml/config.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0"?>
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:noNamespaceSchemaLocation="urn:magento:module:Magento_Store:etc/config.xsd">
+    <default>
+        <reports>
+            <options>
+                <advanced_contacts_enabled>1</advanced_contacts_enabled>
+            </options>
+        </reports>
+    </default>
+</config>
