+++ app/code/Study/ImprovedContact/etc/adminhtml/routes.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0"?>
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:noNamespaceSchemaLocation="urn:magento:framework:App/etc/routes.xsd">
+    <router id="admin">
+        <route id="improvedcontact" frontName="improvedcontact">
+            <module name="Study_ImprovedContact"/>
+        </route>
+    </router>
+</config>
