+++ app/code/Study/ImprovedContact/etc/module.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0"?>
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:noNamespaceSchemaLocation="urn:magento:framework:Module/etc/module.xsd">
+    <module name="Study_ImprovedContact" >
+        <sequence>
+            <module name="Magento_Contact" />
+        </sequence>
+    </module>
+</config>
