+++ app/code/Study/ImprovedContact/registration.php
@@ -1,0 +1,8 @@
+<?php
+use Magento\Framework\Component\ComponentRegistrar;
+
+ComponentRegistrar::register(
+    ComponentRegistrar::MODULE,
+    'Study_ImprovedContact',
+    __DIR__
+);
