+++ app/code/Study/ImprovedContact/view/adminhtml/layout/improvedcontact_index_index.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0"?>
+<page xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+      xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
+    <update handle="styles"/>
+    <body>
+        <referenceContainer name="content">
+            <uiComponent name="improvedcontact_form_listing"/>
+        </referenceContainer>
+    </body>
+</page>
